backend/pyproject.toml
[project]
name = "agentloopr-backend"
version = "0.1.0"
description = "AgentLoopr backend service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "dramatiq[redis]>=1.18.0,<2.0.0",
  "redis>=6.2.0,<7.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "alembic>=1.14.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "PyJWT>=2.10.1,<3.0.0",
  "pwdlib[argon2]>=0.2.1,<1.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "langgraph>=0.2.0,<1.0.0",
  "cairosvg>=2.7.1,<3.0.0",
  "lxml>=5.2.0,<6.0.0",
  "markdown-it-py>=3.0.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.12.0,<1.0.0",
  "mypy>=1.13.0,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
